Another update ... >:/

+++ disassembler.h
@@ -1,6 +1,6 @@
-#ifndef DISASSEMBLER_H
+#ifndef DISASSEMBLER_H // Was for checking out header guards ...
 #define DISASSEMBLER_H
 
-int disassemble(int32_t instruction);
+int disassemble(uint16_t instruction);
 
 #endif
